Solve day 22
I solved part a yesterday, but needed more time to get the 3d part
working. The face mapping and face coordinates are hard coded in my
solution.

The nice thing about the edge mappings are that they are symmetric,
which means that I can write code which verifies that the mappings
are consistent with one another. Doing so fixed 4 issues with the
mappigns I had already hard coded.

To construct the mapping, I also had to cut out a paper cube to
reference.

And finally, I print the first 100 iterations which revealed some
bugs with the wrapping logic itself.