Solve day 19 part 2
This approach is still inherently backtracking, but it now prunes
useless nodes.

The useless nodes identified are nodes where nothing is purchased
or nodes where even if we built a new geode robot every turn, by
the end, we wouldn't improve over the current highest value.

This was enough to solve part b in only 15 seconds on my 5 year
old laptop.